Load data from the SCAR Southern Ocean Diet and Energetics database

Load data from the SCAR Southern Ocean Diet and Energetics database. Data will be fetched from the remote server and optionally cached locally, or fetched from the local cache, depending on the arguments passed. Note that so_isotopes now has a format parameter - see Details below.

so_diet(method = "get", cache_directory = "session",
  refresh_cache = FALSE, public_only = TRUE, verbose = FALSE)

so_dna_diet(method = "get", cache_directory = "session",
  refresh_cache = FALSE, public_only = TRUE, verbose = FALSE)

so_isotopes(method = "get", cache_directory = "session",
  refresh_cache = FALSE, public_only = TRUE, verbose = FALSE,
  format = "wide")

so_energetics(method = "get", cache_directory = "session",
  refresh_cache = FALSE, public_only = TRUE, verbose = FALSE)

so_lipids(method = "get", cache_directory = "session",
  refresh_cache = FALSE, public_only = TRUE, verbose = FALSE)

Arguments

method
string: "get" (fetch the data via a web GET call) or "direct" (direct database connection, for internal AAD use only)

cache_directory
string: (optional) cache the data locally in this directory, so that they can be used offline later. Values can be "session" (a per-session temporary directory will be used, default), "persistent" (the directory returned by rappdirs::user_cache_dir will be used), or a string giving the path to the directory to use. Use NULL for no caching. An attempt will be made to create the cache directory if it does not exist. A warning will be given if a cached copy of the data file exists and is more than 30 days old. Use refresh_cache = TRUE to refresh the cached data if necessary

refresh_cache
logical: if TRUE, and data already exist in the cache_directory, they will be refreshed. If FALSE, the cached data will be used

public_only
logical: only applicable to method "direct"

verbose
logical: show progress messages?

format
string: (for so_isotopes only) if "wide", the data will be formatted with one row per record and multiple measurements (of different isotopes) per row. If format is "mv", the data will be in measurement-value (long) format with multiple rows per original record, split so that each different isotope measurement appears in its own row. See Details for future changes to the default value of this parameter

Value

data.frame

Details

The format parameter was introduced to the so_isotopes function in package version 0.4.0. The current default format is "wide", in which case the data will be formatted with one row per record and multiple measurements (of different isotopes) per row. If format is "mv", the data will be in measurement-value (long) format with multiple rows per original record, split so that each different isotope measurement appears in its own row. Note that "mv" will become the default (and possibly only) option in a later release. Currently, record_id values in measurement-value format are not unique (they follow the record_id values from the wide format). The record_id values in measurement-value format are likely to change in a future release.

References

http://data.aad.gov.au/trophic/

Examples

# NOT RUN {
  ## conventional diet data
  x <- so_diet()
  subset(x, predator_name == "Electrona antarctica")

  ## DNA diet data
  x <- so_dna_diet()
  subset(x, predator_name == "Thalassarche melanophris")

  ## stable isotopes in wide format
  x <- so_isotopes()

  ## stable isotopes in measurement-value (long) format
  x <- so_isotopes(format = "mv")

  ## energetics data
  x <- so_energetics()

  ## lipids data
  x <- so_lipids()
# }